Refactor: Move helpers to external files
